test(spec): split builder tests into accessor, operators and curves; shared material in conftest
